Change priority in UpdateablePriorityQueue from int to long
This allows a larger range of priorities, and is required to implement
softConcurrencyLimits while avoiding major code changes.

+++ presto-main/src/main/java/com/facebook/presto/execution/resourceGroups/UpdateablePriorityQueue.java
@@ -16,7 +16,7 @@ package com.facebook.presto.execution.resourceGroups;
 interface UpdateablePriorityQueue<E>
         extends Iterable<E>
 {
-    boolean addOrUpdate(E element, int priority);
+    boolean addOrUpdate(E element, long priority);
 
     boolean contains(E element);
